Fix xctestrun file detection
I recognized that the iOS version and the simulator version not always
match! E.g. now Xcode shows iOS 18.2 SDK but iPhone 18.3.1 Simulator.
That was why xcodebuild uses the 18.2 SDK version in the xctestrun which
therefore could not be found.

Now the file is searched based on extension and test-plan name. Fingers
crossed that this is more secure. At least the fishy, custom
construction of that file name is now gone…